requirements.txt
eth-brownie~=1.19.0
pycoingecko~=2.2.0
requests~=2.27.1
